Converting Announcements to a class
